+++ CHANGELOG.md
@@ -22,7 +22,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Catch up with FreeBSD Ports:
   - Remove support for `PKGUPGRADE` and friends
 - portclippy, portscan: Report unknown target sources too
-- portscan: default `-p` to `/usr/ports`
+- portscan: default `-p` to `/usr/ports` or the value of `PORTSDIR` in the environment
 - portedit, portfmt: Stop messing with inline comments.  This should let it
   deal better with the commonly used `PATCHFILES+=<commit>.patch # <pr>`
   pattern.